requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
autopep8=1.6.0=pypi_0
azure-cognitiveservices-speech=1.25.0=pypi_0
ca-certificates=2023.01.10=hecd8cb5_0
certifi=2021.10.8=pypi_0
charset-normalizer=2.0.7=pypi_0
click=8.0.3=pypi_0
et-xmlfile=1.1.0=pypi_0
flask=2.0.2=pypi_0
idna=3.3=pypi_0
itsdangerous=2.0.1=pypi_0
jinja2=3.0.2=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.4.2=hecd8cb5_6
markupsafe=2.0.1=pypi_0
ncurses=6.4=hcec6c5f_0
numpy=1.21.3=pypi_0
openai=0.19.0=pypi_0
openpyxl=3.0.9=pypi_0
openssl=1.1.1t=hca72f7f_0
pandas=1.3.4=pypi_0
pandas-stubs=1.2.0.35=pypi_0
pip=22.3.1=py39hecd8cb5_0
pycodestyle=2.8.0=pypi_0
pyobjc=9.0.1=pypi_0
pyobjc-core=9.0.1=pypi_0
pyobjc-framework-accessibility=9.0.1=pypi_0
pyobjc-framework-accounts=9.0.1=pypi_0
pyobjc-framework-addressbook=9.0.1=pypi_0
pyobjc-framework-adservices=9.0.1=pypi_0
pyobjc-framework-adsupport=9.0.1=pypi_0
pyobjc-framework-applescriptkit=9.0.1=pypi_0
pyobjc-framework-applescriptobjc=9.0.1=pypi_0
pyobjc-framework-applicationservices=9.0.1=pypi_0
pyobjc-framework-apptrackingtransparency=9.0.1=pypi_0
pyobjc-framework-audiovideobridging=9.0.1=pypi_0
pyobjc-framework-authenticationservices=9.0.1=pypi_0
pyobjc-framework-automaticassessmentconfiguration=9.0.1=pypi_0
pyobjc-framework-automator=9.0.1=pypi_0
pyobjc-framework-avfoundation=9.0.1=pypi_0
pyobjc-framework-avkit=9.0.1=pypi_0
pyobjc-framework-avrouting=9.0.1=pypi_0
pyobjc-framework-backgroundassets=9.0.1=pypi_0
pyobjc-framework-businesschat=9.0.1=pypi_0
pyobjc-framework-calendarstore=9.0.1=pypi_0
pyobjc-framework-callkit=9.0.1=pypi_0
pyobjc-framework-cfnetwork=9.0.1=pypi_0
pyobjc-framework-classkit=9.0.1=pypi_0
pyobjc-framework-cloudkit=9.0.1=pypi_0
pyobjc-framework-cocoa=9.0.1=pypi_0
pyobjc-framework-collaboration=9.0.1=pypi_0
pyobjc-framework-colorsync=9.0.1=pypi_0
pyobjc-framework-contacts=9.0.1=pypi_0
pyobjc-framework-contactsui=9.0.1=pypi_0
pyobjc-framework-coreaudio=9.0.1=pypi_0
pyobjc-framework-coreaudiokit=9.0.1=pypi_0
pyobjc-framework-corebluetooth=9.0.1=pypi_0
pyobjc-framework-coredata=9.0.1=pypi_0
pyobjc-framework-corehaptics=9.0.1=pypi_0
pyobjc-framework-corelocation=9.0.1=pypi_0
pyobjc-framework-coremedia=9.0.1=pypi_0
pyobjc-framework-coremediaio=9.0.1=pypi_0
pyobjc-framework-coremidi=9.0.1=pypi_0
pyobjc-framework-coreml=9.0.1=pypi_0
pyobjc-framework-coremotion=9.0.1=pypi_0
pyobjc-framework-coreservices=9.0.1=pypi_0
pyobjc-framework-corespotlight=9.0.1=pypi_0
pyobjc-framework-coretext=9.0.1=pypi_0
pyobjc-framework-corewlan=9.0.1=pypi_0
pyobjc-framework-cryptotokenkit=9.0.1=pypi_0
pyobjc-framework-datadetection=9.0.1=pypi_0
pyobjc-framework-devicecheck=9.0.1=pypi_0
pyobjc-framework-dictionaryservices=9.0.1=pypi_0
pyobjc-framework-discrecording=9.0.1=pypi_0
pyobjc-framework-discrecordingui=9.0.1=pypi_0
pyobjc-framework-diskarbitration=9.0.1=pypi_0
pyobjc-framework-dvdplayback=9.0.1=pypi_0
pyobjc-framework-eventkit=9.0.1=pypi_0
pyobjc-framework-exceptionhandling=9.0.1=pypi_0
pyobjc-framework-executionpolicy=9.0.1=pypi_0
pyobjc-framework-extensionkit=9.0.1=pypi_0
pyobjc-framework-externalaccessory=9.0.1=pypi_0
pyobjc-framework-fileprovider=9.0.1=pypi_0
pyobjc-framework-fileproviderui=9.0.1=pypi_0
pyobjc-framework-findersync=9.0.1=pypi_0
pyobjc-framework-fsevents=9.0.1=pypi_0
pyobjc-framework-gamecenter=9.0.1=pypi_0
pyobjc-framework-gamecontroller=9.0.1=pypi_0
pyobjc-framework-gamekit=9.0.1=pypi_0
pyobjc-framework-gameplaykit=9.0.1=pypi_0
pyobjc-framework-healthkit=9.0.1=pypi_0
pyobjc-framework-imagecapturecore=9.0.1=pypi_0
pyobjc-framework-imserviceplugin=9.0.1=pypi_0
pyobjc-framework-inputmethodkit=9.0.1=pypi_0
pyobjc-framework-installerplugins=9.0.1=pypi_0
pyobjc-framework-instantmessage=9.0.1=pypi_0
pyobjc-framework-intents=9.0.1=pypi_0
pyobjc-framework-intentsui=9.0.1=pypi_0
pyobjc-framework-iosurface=9.0.1=pypi_0
pyobjc-framework-ituneslibrary=9.0.1=pypi_0
pyobjc-framework-kernelmanagement=9.0.1=pypi_0
pyobjc-framework-latentsemanticmapping=9.0.1=pypi_0
pyobjc-framework-launchservices=9.0.1=pypi_0
pyobjc-framework-libdispatch=9.0.1=pypi_0
pyobjc-framework-linkpresentation=9.0.1=pypi_0
pyobjc-framework-localauthentication=9.0.1=pypi_0
pyobjc-framework-localauthenticationembeddedui=9.0.1=pypi_0
pyobjc-framework-mailkit=9.0.1=pypi_0
pyobjc-framework-mapkit=9.0.1=pypi_0
pyobjc-framework-mediaaccessibility=9.0.1=pypi_0
pyobjc-framework-medialibrary=9.0.1=pypi_0
pyobjc-framework-mediaplayer=9.0.1=pypi_0
pyobjc-framework-mediatoolbox=9.0.1=pypi_0
pyobjc-framework-metal=9.0.1=pypi_0
pyobjc-framework-metalfx=9.0.1=pypi_0
pyobjc-framework-metalkit=9.0.1=pypi_0
pyobjc-framework-metalperformanceshaders=9.0.1=pypi_0
pyobjc-framework-metalperformanceshadersgraph=9.0.1=pypi_0
pyobjc-framework-metrickit=9.0.1=pypi_0
pyobjc-framework-mlcompute=9.0.1=pypi_0
pyobjc-framework-modelio=9.0.1=pypi_0
pyobjc-framework-multipeerconnectivity=9.0.1=pypi_0
pyobjc-framework-naturallanguage=9.0.1=pypi_0
pyobjc-framework-netfs=9.0.1=pypi_0
pyobjc-framework-network=9.0.1=pypi_0
pyobjc-framework-networkextension=9.0.1=pypi_0
pyobjc-framework-notificationcenter=9.0.1=pypi_0
pyobjc-framework-opendirectory=9.0.1=pypi_0
pyobjc-framework-osakit=9.0.1=pypi_0
pyobjc-framework-oslog=9.0.1=pypi_0
pyobjc-framework-passkit=9.0.1=pypi_0
pyobjc-framework-pencilkit=9.0.1=pypi_0
pyobjc-framework-photos=9.0.1=pypi_0
pyobjc-framework-photosui=9.0.1=pypi_0
pyobjc-framework-preferencepanes=9.0.1=pypi_0
pyobjc-framework-pushkit=9.0.1=pypi_0
pyobjc-framework-quartz=9.0.1=pypi_0
pyobjc-framework-quicklookthumbnailing=9.0.1=pypi_0
pyobjc-framework-replaykit=9.0.1=pypi_0
pyobjc-framework-safariservices=9.0.1=pypi_0
pyobjc-framework-safetykit=9.0.1=pypi_0
pyobjc-framework-scenekit=9.0.1=pypi_0
pyobjc-framework-screencapturekit=9.0.1=pypi_0
pyobjc-framework-screensaver=9.0.1=pypi_0
pyobjc-framework-screentime=9.0.1=pypi_0
pyobjc-framework-scriptingbridge=9.0.1=pypi_0
pyobjc-framework-searchkit=9.0.1=pypi_0
pyobjc-framework-security=9.0.1=pypi_0
pyobjc-framework-securityfoundation=9.0.1=pypi_0
pyobjc-framework-securityinterface=9.0.1=pypi_0
pyobjc-framework-servicemanagement=9.0.1=pypi_0
pyobjc-framework-sharedwithyou=9.0.1=pypi_0
pyobjc-framework-sharedwithyoucore=9.0.1=pypi_0
pyobjc-framework-shazamkit=9.0.1=pypi_0
pyobjc-framework-social=9.0.1=pypi_0
pyobjc-framework-soundanalysis=9.0.1=pypi_0
pyobjc-framework-speech=9.0.1=pypi_0
pyobjc-framework-spritekit=9.0.1=pypi_0
pyobjc-framework-storekit=9.0.1=pypi_0
pyobjc-framework-syncservices=9.0.1=pypi_0
pyobjc-framework-systemconfiguration=9.0.1=pypi_0
pyobjc-framework-systemextensions=9.0.1=pypi_0
pyobjc-framework-threadnetwork=9.0.1=pypi_0
pyobjc-framework-uniformtypeidentifiers=9.0.1=pypi_0
pyobjc-framework-usernotifications=9.0.1=pypi_0
pyobjc-framework-usernotificationsui=9.0.1=pypi_0
pyobjc-framework-videosubscriberaccount=9.0.1=pypi_0
pyobjc-framework-videotoolbox=9.0.1=pypi_0
pyobjc-framework-virtualization=9.0.1=pypi_0
pyobjc-framework-vision=9.0.1=pypi_0
pyobjc-framework-webkit=9.0.1=pypi_0
python=3.9.16=h218abb5_0
python-dateutil=2.8.2=pypi_0
python-dotenv=0.19.2=pypi_0
pyttsx3=2.90=pypi_0
pytz=2021.3=pypi_0
readline=8.2=hca72f7f_0
requests=2.26.0=pypi_0
setuptools=65.6.3=py39hecd8cb5_0
six=1.16.0=pypi_0
sqlite=3.40.1=h880c91c_0
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pypi_0
tqdm=4.62.3=pypi_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.7=pypi_0
werkzeug=2.0.2=pypi_0
wheel=0.38.4=py39hecd8cb5_0
xz=5.2.10=h6c40b1e_1
zlib=1.2.13=h4dc903c_0
